boost: Fixed boost posix_time compilation issue for boost 1.67 - Explicit conversions from floats to longs avoids the failure

+++ host/examples/rx_samples_to_file.cpp
@@ -182,7 +182,7 @@ bool check_locked_sensor(std::vector<std::string> sensor_names, const char* sens
 
     while (true) {
         if ((not first_lock_time.is_not_a_date_time()) and
-                (boost::get_system_time() > (first_lock_time + boost::posix_time::seconds(setup_time))))
+                (boost::get_system_time() > (first_lock_time + boost::posix_time::seconds(int64_t(setup_time)))))
         {
             std::cout << " locked." << std::endl;
             break;
@@ -196,7 +196,7 @@ bool check_locked_sensor(std::vector<std::string> sensor_names, const char* sens
         else {
             first_lock_time = boost::system_time();	//reset to 'not a date time'
 
-            if (boost::get_system_time() > (start + boost::posix_time::seconds(setup_time))){
+            if (boost::get_system_time() > (start + boost::posix_time::seconds(int64_t(setup_time)))){
                 std::cout << std::endl;
                 throw std::runtime_error(str(boost::format("timed out waiting for consecutive locks on sensor \"%s\"") % sensor_name));
             }
@@ -317,7 +317,7 @@ int UHD_SAFE_MAIN(int argc, char *argv[]){
     //set the antenna
     if (vm.count("ant")) usrp->set_rx_antenna(ant);
 
-    boost::this_thread::sleep(boost::posix_time::seconds(setup_time)); //allow for some setup time
+    boost::this_thread::sleep(boost::posix_time::seconds(int64_t(setup_time))); //allow for some setup time
 
     //check Ref and LO Lock detect
     if (not vm.count("skip-lo")){
